separated player class
