+++ level.py
@@ -22,5 +22,7 @@ class Tile(object):
     def __str__(self):
         return "Tile X Coord: {}\nTile Y Coord: {}".format(self.x, self.y)
 
-levelchunk1 = [Tile("Tiles/Grass_Top.png", 0.0, 85,  150),Tile("Tiles/Grass_Top.png", 0.0, 115,  150), Tile("Tiles/Grass_Top.png", 0.0, 130,  150), Tile("Tiles/Grass_Top.png", 0.0, 145,  150), Tile("Tiles/Grass_Top.png", 0.0, 160,  150), Tile("Tiles/Grass_Top.png", 0.0, 100,  150), Tile("Tiles/Grass_Top.png", 0.0, 70,  150), Tile("Tiles/Grass_Top.png", 0.0, 115,  135)]
+levelchunk1 = [Tile("Tiles/Grass_Top.png", 0.0, 85,  150),Tile("Tiles/Grass_Top.png", 0.0, 70,  150), Tile("Tiles/Grass_Top.png", 0.0, 130,  150), \
+    Tile("Tiles/Grass_Top.png", 0.0, 145,  150), Tile("Tiles/Grass_Top.png", 0.0, 160,  150), Tile("Tiles/Grass_Top.png", 0.0, 100,  150),\
+    Tile("Tiles/Grass_Top.png", 0.0, 70,  150), Tile("Tiles/Grass_Top.png", 0.0, 115,  80)]
 tile = Tile("Tiles/Grass_Top.png", 0.0, 100,  150)
