+++ main.py
@@ -13,106 +13,13 @@ DEBUG = False
 import pygame
 import sys
 from level import *
+from player import *
 
 # Allows us to use another folder than the folder this file is located
 sys.path.insert(1, "./Sprites")
 
+#---------------------------------#
 ##########--BEGIN CLASSES--##########
-
-# Class for the player
-class Player(object):
-    def __init__(self, skin = None, height = 0, weight = 0.2, player_number = 0):
-        self.player_number = player_number
-        # Used to determine what sprites to load for the player
-        self.skin = skin
-        self.x = 100
-        self.y = 100
-        self.x_velocity = 0.00
-        self.y_velocity = 0.00
-        # Number is from height of the player sprite in pixles
-        self.height = height
-        self.weight = weight
-    
-    def gravity(self, gravity):
-        p_weight = self.weight
-        if ((player.y_velocity >= VSPEED_CAP) and (player.y_velocity < 0)):
-            if p_weight <= gravity:
-                    p_weight += gravity
-                    
-            else:
-                p_weight = gravity
-            player.y_velocity += (player.weight * p_weight)
-
-            if player.y_velocity >= 0:
-                player.y_velocity = 0.0
-
-        elif((player.y_velocity >= 0.0) and (player.check_fall() ==False)):
-            if p_weight <= gravity:
-                p_weight += gravity
-            else:
-                p_weight = gravity
-
-            player.y_velocity += (player.weight * p_weight)
-            if player.y_velocity < VSPEED_CAP:
-                player.y_velocity = VSPEED_CAP
-
-        else:
-            if player.check_fall() != False:
-                player.y_velocity = 0.0
-                player.y = player.check_fall()
-    
-    def death(self):
-        pass
-
-    def respawn(self):
-        pass
-    
-    def check_jump(self):
-        for i in range(len(levelchunk1)):
-            if (player.y == levelchunk1[i-1].top) and player.x <= levelchunk1[i-1].width + levelchunk1[i-1].x and player.x >=levelchunk1[i-1].x - levelchunk1[i-1].x: 
-                return True
-        return False
-
-    def check_fall(self):
-        for i in range(len(levelchunk1)):
-            if (player.y >= levelchunk1[i-1].top) and player.y <= levelchunk1[i-1].top + 10 and player.x <= levelchunk1[i-1].width + levelchunk1[i-1].x and player.x >=levelchunk1[i-1].x - 16: 
-                return levelchunk1[i-1].top
-        return False
-
-    # Check to see if the player can jump
-    def check_jump(self):
-        for i in range(len(levelchunk1)):
-            if (player.y == levelchunk1[i-1].top) and player.x <= levelchunk1[i-1].width + levelchunk1[i-1].x\
-                and player.x >=levelchunk1[i-1].x - levelchunk1[i-1].x: 
-                return True
-        return False
-
-    # Check to see if the player should have gravity applied
-    def check_fall(self):
-        for i in range(len(levelchunk1)):
-            if (player.y >= levelchunk1[i-1].top) and player.y <= levelchunk1[i-1].top + 10\
-                and player.x <= levelchunk1[i-1].width + levelchunk1[i-1].x and player.x >=levelchunk1[i-1].x - 15: 
-                return levelchunk1[i-1].top
-        return False
-
-    def calculatePosition(self):
-        # Make it so the player wraps around on the left and right (if enabled)
-        if (wrap_around):
-            if (((player.x >= WIDTH - 10) and (player.x <= WIDTH)) and player.x_velocity >= 0):
-                player.x = 11
-            elif ((player.x >= 0) and (player.x <= 10)) and (player.x_velocity <= 0):
-                player.x = WIDTH - 11
-        
-        # Calculate the players next position using their coordinates
-        # (This will probably be improved in the future)
-        self.x += self.x_velocity
-        self.y += self.y_velocity
-
-    
-    # Print stats of the player when called
-    def __str__(self):
-        return "Player X Velocity: {}\nPlayer Y Velocity: {}\nPlayer X: {}\nPlayer Y: {}"\
-            .format(player.x_velocity, player.y_velocity, player.x, player.y)
   
 ##########--END CLASSES--##########
 #---------------------------------#
@@ -121,7 +28,7 @@ class Player(object):
 def check_colision():
     for i in range(len(levelchunk1)):
         if (player.x >= levelchunk1[i-1].left -11) and player.x <= levelchunk1[i-1].right + 9 and player.y -1 > levelchunk1[i-1].top and player.y - 1 <=levelchunk1[i-1].y: 
-            if player.x - levelchunk1[i-1].left-11 < levelchunk1[i-1].right - player.x + 4:
+            if player.x - levelchunk1[i-1].left-11 < levelchunk1[i-1].right - player.x:
                 return levelchunk1[i-1].left -11
             else:
                 return levelchunk1[i-1].right + 9
@@ -139,17 +46,11 @@ clock = pygame.time.Clock()
 # Make the window title
 pygame.display.set_caption("Mario vs Luigi")
 
-# Create a player
-player = Player("Sprites/idle.png",20)
 # Define some constants
-SIZE = WIDTH, HEIGHT = 320, 240
 BLACK = (0, 0, 0)
 
 # Setup the screen and other display stuff
 screen = pygame.display.set_mode(SIZE)
-
-# Define some settings variables
-wrap_around = True
 
 # Define some in game constants (used for the Physics "engine")
 SPEED_CAP = 8.0
@@ -160,9 +61,12 @@ V_ACCELERATION = 0.1
 GRAVITY = 2.5
 
 # Define misc variables
-xc = [85,115,130,145,160,100,70,115]
-yc = [150,150,150,150,150,150,150,135]
 frame = 1
+
+# Create a player
+player = Player("Sprites/idle.png",18)
+playerL = Player("Sprites/idle.png",20)
+players = [player,playerL]
 
 # Define player controls
 up = pygame.K_UP
@@ -248,7 +152,7 @@ while True:
     #Render the screen
     screen.fill(BLACK)
     for i in range(len(levelchunk1)):
-        screen.blit(pygame.image.load(levelchunk1[i-1].tile_image), [xc[i-1], yc[i-1]])
+        screen.blit(pygame.image.load(levelchunk1[i-1].tile_image), [levelchunk1[i-1].x, levelchunk1[i-1].y])
     screen.blit(pygame.image.load(player.skin), [player.x, player.y - player.height])
     pygame.display.flip()
